requirements.txt
pyserial>=3.5
google-genai>=2.0.0
Pillow>=10.0.0
